requirements.txt
pyserial==2.5
evdev==0.4.1
